Switched imgui to docking branch, using their GLFW/OpenGL implementation and switched Hazel to static lib

+++ Hazel/src/Hazel/Events/MouseEvent.h
@@ -50,7 +50,7 @@ namespace Hazel {
 	class HAZEL_API MouseButtonEvent : public Event
 	{
 	protected:
-		float m_Button;
+		int m_Button;
 
 		MouseButtonEvent(int button)
 			: m_Button(button) {}
